add authorization for RuleController

+++ services/room-scheduler/src/main/java/roomscheduler/controllers/RuleController.java
@@ -5,13 +5,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 import roomscheduler.entities.Rule;
 import roomscheduler.repositories.RuleRepository;
 
@@ -31,9 +25,13 @@ public class RuleController {
     }
 
     @PostMapping(path = "/createRule") // Map ONLY POST Requests
-    public @ResponseBody String addNewRule(@RequestBody Rule r) throws IOException {
-        ruleRepository.saveAndFlush(r);
-        return "Saved rule";
+    public @ResponseBody String addNewRule(@RequestHeader("Authorization") String token,
+                                           @RequestBody Rule r) throws IOException {
+        if (Authorization.authorize(token, "Teacher")) {
+            ruleRepository.saveAndFlush(r);
+            return "Saved rule";
+        } else throw new RuntimeException("You do not have the privilege to perform " +
+                "this action.");
     }
 
     /**
@@ -42,8 +40,11 @@ public class RuleController {
      * @return Array List with all rules
      */
     @GetMapping(path = "/allRules")
-    public @ResponseBody List<Rule> getAllRules() {
-        return ruleRepository.findAll();
+    public @ResponseBody List<Rule> getAllRules(@RequestHeader("Authorization") String token) {
+        if (Authorization.authorize(token, "Student")) {
+            return ruleRepository.findAll();
+        } else throw new RuntimeException("You do not have the privilege to perform " +
+                "this action.");
     }
 
     /**
@@ -53,16 +54,20 @@ public class RuleController {
      * @return message
      */
     @PutMapping(path = "/modifyRule")
-    public @ResponseBody String editRule(@RequestBody Rule r) {
-        Optional<Rule> rule = ruleRepository.findById(r.getIdrules());
-        if (rule.isPresent()) {
-            rule.get().setName(r.getName());
-            rule.get().setValue(r.getValue());
-            ruleRepository.saveAndFlush(rule.get());
-            return "Changed Rule";
-        } else {
-            throw new RuntimeException("Rule not found for the id " + r.getIdrules());
-        }
+    public @ResponseBody String editRule(@RequestHeader("Authorization") String token,
+                                         @RequestBody Rule r) {
+        if (Authorization.authorize(token, "Teacher")) {
+            Optional<Rule> rule = ruleRepository.findById(r.getIdrules());
+            if (rule.isPresent()) {
+                rule.get().setName(r.getName());
+                rule.get().setValue(r.getValue());
+                ruleRepository.saveAndFlush(rule.get());
+                return "Changed Rule";
+            } else {
+                throw new RuntimeException("Rule not found for the id " + r.getIdrules());
+            }
+        } else throw new RuntimeException("You do not have the privilege to perform " +
+                "this action.");
     }
 
     /**
@@ -72,14 +77,18 @@ public class RuleController {
      * @return message
      */
     @DeleteMapping(path = "/deleteRule/{id}")
-    public @ResponseBody String deleteRule(@PathVariable int id) {
-        Optional<Rule> rule = ruleRepository.findById(id);
-        if (rule.isPresent()) {
-            ruleRepository.deleteById(id);
-            return "Deleted Rule";
-        } else {
-            throw new RuntimeException("Rule not found for the id " + id);
-        }
+    public @ResponseBody String deleteRule(@RequestHeader("Authorization") String token,
+                                           @PathVariable int id) {
+        if (Authorization.authorize(token, "Teacher")) {
+            Optional<Rule> rule = ruleRepository.findById(id);
+            if (rule.isPresent()) {
+                ruleRepository.deleteById(id);
+                return "Deleted Rule";
+            } else {
+                throw new RuntimeException("Rule not found for the id " + id);
+            }
+        } else throw new RuntimeException("You do not have the privilege to perform " +
+                "this action.");
     }
 
 }
